feature: Optimized pubsub api demo (#594)
